dia 04 de febrero
canales

+++ src/es/santiagoayalabustamante/protectfx/Flecha.java
@@ -16,10 +16,10 @@ import javafx.scene.shape.Rectangle;
  */
 public class Flecha extends Group {
     Polygon triangFlecha;
-    int posX = 0;
-    int posY = 0;
-    int velX = 0;
-    int velY = 0;
+    double posX = 0;
+    double posY = 0;
+    double velX = 0;
+    double velY = 0;
     
     public Flecha() {
         Rectangle rectangFlecha = new Rectangle(0, 0, 15, 10);
